requirements.txt
yfinance
pandas
python-dotenv
alpaca-py
